requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
Jinja2==3.0.3
Markdown==3.3.7
MarkupSafe==2.0.1
Pillow==8.4.0
SQLAlchemy==1.4.37
Werkzeug==2.0.3
absl-py==1.1.0
cachetools==4.2.4
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.0.4
dataclasses==0.8
google-auth==2.6.6
google-auth-oauthlib==0.4.6
greenlet==1.1.2
grpcio==1.46.3
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.8.3
importlib-resources==5.4.0
itsdangerous==2.0.1
numpy==1.19.5
oauthlib==3.2.0
pandas==1.1.5
pip==21.3.1
protobuf==3.19.4
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.5.4
setuptools==59.6.0
six==1.16.0
torch==1.8.1+cpu
torchvision==0.9.1+cpu
tqdm==4.64.0
typing-extensions==4.1.1
urllib3==1.26.9
wheel==0.37.1
zipp==3.6.0